# swagger-client/src/main/resources/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This API is responsible to handle pet data.",
		"version": "1.0",
		"title": "Petstore API",
		"termsOfService": "Terms of service",
		"contact": {
			"name": "diegosanteri",
			"url": "www.example.com",
			"email": "diegosanteri@gmail.com"
		},
		"license": {
			"name": "License of API",
			"url": "MTI"
		}
	},
	"host": "localhost:8080",
	"basePath": "/",
	"tags": [{
		"name": "pets",
		"description": "Pet Controller"
	}],
	"schemes": ["http"],
	"paths": {
		"/api/v1/pets": {
			"get": {
				"tags": ["pets"],
				"summary": "findAllPets",
				"operationId": "findAllPetsUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/PetModel"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"post": {
				"tags": ["pets"],
				"summary": "addPet",
				"operationId": "addPetUsingPOST",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"in": "body",
					"name": "pet",
					"description": "pet",
					"required": true,
					"schema": {
						"$ref": "#/definitions/PetModel"
					}
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/PetModel"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/v1/pets/{id}": {
			"get": {
				"tags": ["pets"],
				"summary": "findPetByID",
				"operationId": "findPetByIDUsingGET",
				"consumes": ["application/json"],
				"produces": ["*/*"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "id",
					"required": true,
					"type": "integer",
					"format": "int64"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/PetModel"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"PetModel": {
			"type": "object",
			"properties": {
				"age": {
					"type": "integer",
					"format": "int32"
				},
				"animal": {
					"type": "string",
					"enum": ["DOG", "CAT", "HAMSTER", "BIRD"]
				},
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"owner_name": {
					"type": "string"
				},
				"pet_name": {
					"type": "string"
				}
			}
		}
	}
}